Make math definitions work with mathjax and LaTeX
LaTeX has a preamble, in which we can define new commands.
Mathjax don have this and requieres to add a preamble to every page.
If we use \newcommand in this preamble, LaTeX will complain, that the
command is already defined. You can use \providecommand instead of
\newcommand to fix this, but mathjax does not work with \providecommand,
see https://github.com/mathjax/MathJax/issues/1344.
This is solved by switching to \def for the command definitions.